add ftan support and use idkeys for get-requests
update version to 3.7.15
this is the major part for #16

I have done a little bit of testing, but by far not covered all changes.
I have read again the whole diff of the commit and corrected what I have
spotted during that review, but an extensive testing should be carried
out anyway before we bump the version to the final 4.0.0 release.