handle IDKEYs for recent WB versions (#44)
starting with WB 2.10 IDKEYs can not be used several times, the switch
to enable re-use them for ajax applications does not work anymore.
Admittedly the possibility to reuse them poses a problem, but if we
can not reuse them e.g. for uploads or drag and drop interactions,
IDKEYs are not suitable for use in this scenario, so we have to fallback
on the cleartext id values, at least on latest WB versions.
For WB 2.8.3 and all WBCE versions however we can use IDKEYs in this context.